scan/requirements.txt
google-cloud-bigquery==3.2.0
google-cloud-pubsub==2.12.0